:recycle: rename commandhandler package to command executer
