+++ server_handler/internal/commandexecuter/command_executer.go
@@ -1,4 +1,4 @@
-package commandhandler
+package commandexecuter
 
 import (
 	"context"
